+++ src/libtrx/game/ui/dialogs/base_passport.c
@@ -14,8 +14,8 @@ static int32_t M_GetVisibleRows(void)
     if (TR_VERSION == 2) {
         return 10;
     } else {
-        const int32_t res_h =
-            Scaler_CalcInverse(Viewport_GetHeight(), SCALER_TARGET_TEXT);
+        const int32_t res_h = Scaler_CalcInverse(
+            Viewport_GetHeight(VIEWPORT_GAME), SCALER_TARGET_TEXT);
         if (res_h <= 240) {
             return 5;
         } else if (res_h <= 384) {